+++ src/azure-blob/handlers.js
@@ -1,3 +1,20 @@
+const mysql = require('mysql2/promise');
+const fs = require('fs');
+
+const { BlobServiceClient, StorageSharedKeyCredential } = require('@azure/storage-blob');
+
+const sharedKeyCredential = new StorageSharedKeyCredential(
+    process.env.AZURE_STORAGE_ACCOUNT_NAME,
+    process.env.AZURE_STORAGE_ACCOUNT_KEY
+);
+
+const blobServiceClient = new BlobServiceClient(
+    `https://${process.env.AZURE_STORAGE_ACCOUNT_NAME}.blob.core.windows.net`,
+    sharedKeyCredential
+);
+
+const containerClient = blobServiceClient.getContainerClient(process.env.AZURE_CONTAINER_NAME);
+
 async function uploadFile(req, res) {
     const fileName = req.body.note;
     if (!fileName) {
@@ -73,4 +90,4 @@ module.exports = {
     uploadFile,
     getFiles,
     deleteFile
-  };
+};
